+++ arize/pandas/tracing/validation/spans/spans_validation.py
@@ -13,12 +13,12 @@ from arize.pandas.validation import errors as err
 
 def validate_argument_types(
     spans_dataframe: pd.DataFrame,
-    model_id: str,
+    project_name: str,
     dt_fmt: str,
     model_version: Optional[str] = None,
 ) -> List[err.ValidationError]:
     checks = chain(
-        common_arg_validation._check_field_convertible_to_str(model_id, model_version),
+        common_arg_validation._check_field_convertible_to_str(project_name, model_version),
         common_arg_validation._check_dataframe_type(spans_dataframe),
         common_arg_validation._check_datetime_format_type(dt_fmt),
     )
@@ -45,12 +45,12 @@ def validate_dataframe_form(
 
 def validate_values(
     spans_dataframe: pd.DataFrame,
-    model_id: str,
+    project_name: str,
     model_version: Optional[str] = None,
 ) -> List[err.ValidationError]:
     checks = chain(
         # Common
-        common_value_validation._check_invalid_model_id(model_id),
+        common_value_validation._check_invalid_project_name(project_name),
         common_value_validation._check_invalid_model_version(model_version),
         # Spans specific
         value_validation._check_span_root_field_values(spans_dataframe),